move IdBlender to common
